Additional changes during code review
