+++ editor/src/document/document_message.rs
@@ -1,6 +1,5 @@
 use super::layer_panel::LayerMetadata;
 use super::utility_types::{AlignAggregate, AlignAxis, FlipAxis};
-use super::{ArtboardMessage, MovementMessage, TransformLayerMessage};
 use crate::message_prelude::*;
 
 use graphene::layers::blend_mode::BlendMode;
