+++ editor/src/viewport_tools/tools/mod.rs
@@ -7,6 +7,6 @@ pub mod navigate;
 pub mod path;
 pub mod pen;
 pub mod rectangle;
-pub mod resize;
 pub mod select;
 pub mod shape;
+pub mod shared;
