+++ editor/src/viewport_tools/tools/select.rs
@@ -9,8 +9,9 @@ use crate::misc::{HintData, HintGroup, HintInfo, KeysGroup};
 use crate::viewport_tools::snapping::SnapHandler;
 use crate::viewport_tools::tool::{DocumentToolData, Fsm, ToolActionHandlerData};
 
+use graphene::intersection::Quad;
 use graphene::layers::style::{self, Fill, Stroke};
-use graphene::{Operation, Quad};
+use graphene::Operation;
 
 use glam::{DAffine2, DVec2};
 use serde::{Deserialize, Serialize};
