+++ editor/src/viewport_tools/tools/shared/mod.rs
@@ -1,0 +1,1 @@
+pub mod resize;
