+++ graphene/src/error.rs
@@ -1,0 +1,12 @@
+use super::LayerId;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum DocumentError {
+	LayerNotFound(Vec<LayerId>),
+	InvalidPath,
+	IndexOutOfBounds,
+	NotAFolder,
+	NonReorderableSelection,
+	NotAShape,
+	InvalidFile(String),
+}
